Extract Windows1258 to its own file
